Allows a length to have no unit.
Feliz.transform.translate is an example where this change is useful. The
function accepts `ICssUnit * ICssUnit`. However, the SVG spec does not
expect a unit for the numbers given in the translate part of a
transformation. This allows the api to remain the same (no sudden change
on what is expected), but outputs what is need to have transformations work.