prop.transform altered to be svg friendly
When using a list of [transform.*] for prop.transform, the units px and
deg are removed. Those two are chosed because when using a raw int or float
for the translate and rotate transforms, 'px' and 'deg' are added
respevitvely. Svg does not use units (the former is always user space units
while the ater is always degrees). However, style.transform does use units,
so the removal does not happen for that use case.

This also reverts some of "Changed transform overrides to not add unit on
raw numbers", which is commit a41a6326df27d22e3026683dc43a320611be0090.

This is primarily a non-breaking change, but should be noted that when
using svg, it is better to use ints and floats rather than ICssUnits for
transforms.